Use classes to better handle tests
Because of recent changes from the Octokit library (relying on Node internal fetch), we can't use `nock` anymore during tests. We must now pass a custom `fetch` to the Octokit class to mock request.
To do that, we can no longer only use `handler` with all the code inside it, but use classes. It's also more clean :)
