+++ functions/fixup.js
@@ -1,5 +1,6 @@
-import { Octokit } from '@octokit/rest'
-import { updateStatus, validateWebhook } from './utils/github'
+import { FixupHandler } from './classes/FixupHandler'
+
+const fixup = new FixupHandler(process.env.GITHUB_TOKEN, process.env.NAMESPACE)
 
 export async function handler(event, context, callback) {
   if (event.headers?.['content-type'] === 'application/x-www-form-urlencoded') {
@@ -9,81 +10,5 @@ export async function handler(event, context, callback) {
     })
   }
 
-  let response
-  const githubClient = new Octokit({ auth: process.env.GITHUB_TOKEN })
-  const payload = {
-    success: {
-      state: 'success',
-      description: 'No fixup commits in history',
-      context: `${process.env.NAMESPACE} - Fixup check`,
-    },
-    failure: {
-      state: 'failure',
-      description: 'Fixup commits in history, please squash them!',
-      context: `${process.env.NAMESPACE} - Fixup check`,
-    },
-  }
-
-  const body = JSON.parse(event.body)
-
-  // when creating the webhook
-  if (body?.hook) {
-    try {
-      const message = validateWebhook(body)
-
-      console.log(message)
-
-      return callback(null, {
-        statusCode: 200,
-        body: message,
-      })
-    } catch (e) {
-      console.log(e.message)
-
-      return callback(null, {
-        statusCode: 500,
-        body: e.message,
-      })
-    }
-  }
-
-  if (!body?.pull_request) {
-    response = {
-      statusCode: 500,
-      body: 'Event is not a Pull Request',
-    }
-
-    return callback(null, response)
-  }
-
-  console.log(`Working on repo ${body.repository.full_name} for PR #${body.pull_request.number}`)
-
-  const compare = await githubClient.rest.repos.compareCommitsWithBasehead({
-    owner: body.repository.owner.login,
-    repo: body.repository.name,
-    basehead: `${body.pull_request.base.sha}...${body.pull_request.head.sha}`,
-  })
-
-  // loop through PR labels to see if we found one which should block the PR
-  const validation = compare.data.commits.every(({ commit, parents }) => {
-    const isMerge = parents && parents.length > 1
-
-    if (isMerge) {
-      return true
-    }
-
-    const match = /^fixup! .*$/im.exec(commit.message)
-
-    if (match === null) {
-      return true
-    }
-
-    console.log(`Fixup commit found: "${commit.message}"`)
-
-    return false
-  })
-
-  response = await updateStatus(githubClient, body, validation ? payload.success : payload.failure)
-
-  return callback(null, response)
+  return fixup.handle(JSON.parse(event.body), callback)
 }
