+++ package.json
@@ -2,7 +2,7 @@
   "name": "serverless-github-check",
   "main": "handler.js",
   "scripts": {
-    "test": "NODE_OPTIONS=--no-experimental-fetch jest --silent",
+    "test": "jest --silent",
     "lint": "eslint functions/ tests/"
   },
   "author": "20 Minutes <web-tech@20minutes.fr>",
@@ -22,8 +22,8 @@
     "eslint-plugin-jsx-a11y": "^6.7.1",
     "eslint-plugin-prettier": "^5.0.0",
     "eslint-plugin-react": "^7.33.2",
+    "fetch-mock": "^9.11.0",
     "jest": "^29.7.0",
-    "nock": "^13.3.3",
     "prettier": "^3.0.3",
     "regenerator-runtime": "^0.14.0",
     "serverless-offline": "^13.1.1",
@@ -33,8 +33,8 @@
   },
   "dependencies": {
     "@babel/runtime": "^7.23.1",
-    "@octokit/graphql": "^5.0.6",
-    "@octokit/rest": "^19.0.13",
+    "@octokit/graphql": "^7.0.2",
+    "@octokit/rest": "^20.0.2",
     "semver-diff": "3.1.1",
     "source-map-support": "^0.5.21"
   },
